+++ tests/NetteDI.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 require __DIR__ . "/../vendor/autoload.php";
 
-use MyTester\Bridges\NetteDI\TestsRunner;
+use MyTester\Tester;
 use Nette\Configurator;
 
 $configurator = new Configurator();
@@ -12,8 +12,7 @@ Nette\Utils\FileSystem::createDir(__DIR__ . "/temp");
 $configurator->setTempDirectory(__DIR__ . "/temp");
 $configurator->addConfig(__DIR__ . "/config.neon");
 $container = $configurator->createContainer();
-/** @var TestsRunner $runner */
-$runner = $container->getByType(TestsRunner::class);
-$result = $runner->execute();
-exit((int) $result);
+/** @var Tester $runner */
+$runner = $container->getByType(Tester::class);
+$runner->execute();
 ?>